requirements.txt
asgiref==3.7.2
Django==4.2.4
django-phonenumber-field==7.1.0
django-phonenumbers==1.0.1
gunicorn==21.2.0
packaging==23.1
phonenumbers==8.13.18
sqlparse==0.4.4
typing_extensions==4.7.1
whitenoise==6.5.0
